services/message-service/requirements.txt
cassandra-driver
nats-py
grpcio
grpcio-tools
pydantic-settings